add base svg frame class and extend it
