+++ apps/src/blockly/addons/blockSvgUnused.js
@@ -2,7 +2,15 @@ import msg from '@cdo/locale';
 import color from '@cdo/apps/util/color';
 import BlockSvgFrame from './blockSvgFrame.js';
 
+/**
+ * Represents an SVG frame specifically designed for unused blocks.
+ * These frames use different style rules and are removed when the block is touched.
+ */
 export default class BlockSvgUnused extends BlockSvgFrame {
+  /**
+   * Constructs an svg frame for an unused block.
+   * @param {Element} block - The unused block element associated with the frame.
+   */
   constructor(block) {
     var className = 'blocklyUnusedFrame';
     var text = msg.unusedCode();
@@ -11,19 +19,13 @@ export default class BlockSvgUnused extends BlockSvgFrame {
     var headerColor = color.lighter_gray;
     super(block, text, className, textColor, headerColor);
   }
-
-  dispose() {
-    this.frameGroup_.remove();
-    this.frameGroup_ = undefined;
-    this.frameClipRect_ = undefined;
-    this.frameBase_ = undefined;
-    this.frameHeader_ = undefined;
-    this.frameText_ = undefined;
-  }
 }
 
-// Added as a change listener in the wrapper.
-// When a block is clicked, dragged or deleted, we remove any "Unused block" frame.
+/**
+ * Event handler for block click, drag, or delete events.
+ * Removes the "Unused block" frame when a block is touched.
+ * @param {Object} event - The Blockly event object.
+ */
 export function onBlockClickDragDelete(event) {
   if (
     event.type === Blockly.Events.CLICK ||
