+++ apps/src/blockly/customBlocks/googleBlockly/behaviorBlocks.js
@@ -134,7 +134,7 @@ GoogleBlockly.Extensions.register('behaviors_block_frame', function () {
 
     this.setOnChange(function () {
       if (!this.isInFlyout) {
-        this.functionalSvg_.render(this.svgGroup_, this.RTL);
+        this.functionalSvg_.render();
       }
     });
   }
